streamlit_app/requirements.txt
streamlit>=1.40.0
openai>=1.57.0
pypdf>=5.1.0
python-docx>=1.1.2
